Fixed fatal error in test if PDO extension was not loaded

+++ tests/phpunit/ImboIntegrationTest/EventListener/ImageVariations/Database/DoctrineTest.php
@@ -71,8 +71,10 @@ class DoctrineTest extends DatabaseTests {
      * Clean up after each run
      */
     public function tearDown() {
-        $this->pdo->query('DROP TABLE IF EXISTS imagevariations');
-        $this->pdo = null;
+        if ($this->pdo) {
+            $this->pdo->query('DROP TABLE IF EXISTS imagevariations');
+            $this->pdo = null;
+        }
 
         parent::tearDown();
     }
